+++ src/YandexWeatherApi/Models/ForecastModels/PartShort.cs
@@ -16,9 +16,9 @@ public class PartShort
 
     [JsonPropertyName("wind_dir")] public string WindDir { get; set; }
 
-    [JsonPropertyName("pressure_mm")] public int PressureMm { get; set; }
+    [JsonPropertyName("pressure_mm")] public float PressureMm { get; set; }
 
-    [JsonPropertyName("pressure_pa")] public int PressurePa { get; set; }
+    [JsonPropertyName("pressure_pa")] public float PressurePa { get; set; }
 
     [JsonPropertyName("humidity")] public int Humidity { get; set; }
 
@@ -26,17 +26,17 @@ public class PartShort
 
     [JsonPropertyName("soil_moisture")] public double SoilMoisture { get; set; }
 
-    [JsonPropertyName("prec_mm")] public int PrecMm { get; set; }
+    [JsonPropertyName("prec_mm")] public float PrecMm { get; set; }
 
     [JsonPropertyName("prec_prob")] public int PrecProb { get; set; }
 
     [JsonPropertyName("prec_period")] public int PrecPeriod { get; set; }
 
-    [JsonPropertyName("cloudness")] public int Cloudness { get; set; }
+    [JsonPropertyName("cloudness")] public float Cloudness { get; set; }
 
     [JsonPropertyName("prec_type")] public int PrecType { get; set; }
 
-    [JsonPropertyName("prec_strength")] public int PrecStrength { get; set; }
+    [JsonPropertyName("prec_strength")] public float PrecStrength { get; set; }
 
     [JsonPropertyName("icon")] public string Icon { get; set; }
